backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
psycopg[binary]>=3.1,<4.0
django-environ>=0.11,<1.0
Pillow>=10.0,<11.0
google-auth>=2.28,<3.0
social-auth-app-django>=5.0,<6.0
